+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added the user profiles feature interface, and it is accessible through "Countly::instance()::userProfile()" call.
 
 * Fixed a bug where setting custom user properties would not work.
+* Fixed a bug where setting organization of the user would not work.
 
 * Deprecated "Countly::backendMode()" call, use "Countly::backendM" instead via "instance()" call.
 * The following methods are deprecated from the "UserEditor" interface: